+++ packages/forge2d/lib/src/callbacks/contact_listener.dart
@@ -11,10 +11,10 @@ import '../../forge2d.dart';
 /// @warning You cannot create/destroy Forge2D entities inside these callbacks.
 abstract class ContactListener {
   /// Called when two fixtures begin to touch.
-  void beginContact(Contact contact);
+  void beginContact(Contact contact) {}
 
   /// Called when two fixtures cease to touch.
-  void endContact(Contact contact);
+  void endContact(Contact contact) {}
 
   /// This is called after a contact is updated. This allows you to inspect a
   /// contact before it goes to the solver. If you are careful, you can modify the
@@ -28,7 +28,7 @@ abstract class ContactListener {
   /// the next step.
   /// Note: the oldManifold parameter is pooled, so it will be the same object for every callback
   /// for each thread.
-  void preSolve(Contact contact, Manifold oldManifold);
+  void preSolve(Contact contact, Manifold oldManifold) {}
 
   /// This lets you inspect a contact after the solver is finished. This is useful
   /// for inspecting impulses.
@@ -38,5 +38,5 @@ abstract class ContactListener {
   /// Note: this is only called for contacts that are touching, solid, and awake.
   /// @param contact
   /// @param impulse this is usually a pooled variable, so it will be modified after this call
-  void postSolve(Contact contact, ContactImpulse impulse);
+  void postSolve(Contact contact, ContactImpulse impulse) {}
 }