+++ packages/forge2d/lib/src/dynamics/joints/joint.dart
@@ -3,43 +3,10 @@ import '../../../forge2d.dart';
 /// The base joint class. Joints are used to constrain two bodies together in various fashions. Some
 /// joints also feature limits and motors.
 abstract class Joint {
-  static T create<T extends Joint>(World world, JointDef def) {
-    if (T == ConstantVolumeJoint || def.type == JointType.constantVolume) {
-      return ConstantVolumeJoint(world, def as ConstantVolumeJointDef) as T;
-    } else if (T == DistanceJoint || def.type == JointType.distance) {
-      return DistanceJoint(def as DistanceJointDef) as T;
-    } else if (T == FrictionJoint || def.type == JointType.friction) {
-      return FrictionJoint(def as FrictionJointDef) as T;
-    } else if (T == GearJoint || def.type == JointType.gear) {
-      return GearJoint(def as GearJointDef) as T;
-    } else if (T == MotorJoint || def.type == JointType.motor) {
-      return MotorJoint(def as MotorJointDef) as T;
-    } else if (T == MouseJoint || def.type == JointType.mouse) {
-      return MouseJoint(def as MouseJointDef) as T;
-    } else if (T == PrismaticJoint || def.type == JointType.prismatic) {
-      return PrismaticJoint(def as PrismaticJointDef) as T;
-    } else if (T == PulleyJoint || def.type == JointType.pulley) {
-      return PulleyJoint(def as PulleyJointDef) as T;
-    } else if (T == RevoluteJoint || def.type == JointType.revolute) {
-      return RevoluteJoint(def as RevoluteJointDef) as T;
-    } else if (T == RopeJoint || def.type == JointType.rope) {
-      return RopeJoint(def as RopeJointDef) as T;
-    } else if (T == WeldJoint || def.type == JointType.weld) {
-      return WeldJoint(def as WeldJointDef) as T;
-    } else if (T == WheelJoint || def.type == JointType.wheel) {
-      return WheelJoint(def as WheelJointDef) as T;
-    } else {
-      throw ArgumentError(
-        'Invalid joint type $T with invalid joint definition ${def.type}',
-      );
-    }
-  }
-
   static void destroy(Joint joint) {
     joint.destructor();
   }
 
-  final JointType _type;
   late Body bodyA;
   late Body bodyB;
 
@@ -52,18 +19,12 @@ abstract class Joint {
   Joint(JointDef def)
       : assert(def.bodyA != def.bodyB),
         localAnchorA = def.localAnchorA,
-        localAnchorB = def.localAnchorB,
-        _type = def.type {
+        localAnchorB = def.localAnchorB {
     bodyA = def.bodyA;
     bodyB = def.bodyB;
     _collideConnected = def.collideConnected;
     islandFlag = false;
   }
-
-  /// get the type of the concrete joint.
-  ///
-  /// @return
-  JointType get type => _type;
 
   /// Whether the body is connected to the joint
   bool containsBody(Body body) => body == bodyA || body == bodyB;
@@ -119,46 +80,8 @@ abstract class Joint {
   /// Override to handle destruction of joint
   void destructor() {}
 
-  final Color3i _color = Color3i.zero();
+  /// Color used to [render].
+  final renderColor = Color3i.zero()..setFromRGBd(0.5, 0.8, 0.8);
 
-  void render(DebugDraw debugDraw) {
-    final xf1 = bodyA.transform;
-    final xf2 = bodyB.transform;
-    final x1 = xf1.p;
-    final x2 = xf2.p;
-    final p1 = anchorA;
-    final p2 = anchorB;
-
-    _color.setFromRGBd(0.5, 0.8, 0.8);
-
-    switch (type) {
-      case JointType.distance:
-        debugDraw.drawSegment(p1, p2, _color);
-        break;
-
-      case JointType.pulley:
-        {
-          final pulley = this as PulleyJoint;
-          final s1 = pulley.getGroundAnchorA();
-          final s2 = pulley.getGroundAnchorB();
-          debugDraw.drawSegment(s1, p1, _color);
-          debugDraw.drawSegment(s2, p2, _color);
-          debugDraw.drawSegment(s1, s2, _color);
-        }
-        break;
-
-      case JointType.friction:
-        debugDraw.drawSegment(x1, x2, _color);
-        break;
-
-      case JointType.constantVolume:
-      case JointType.mouse:
-        // don't draw this
-        break;
-      default:
-        debugDraw.drawSegment(x1, p1, _color);
-        debugDraw.drawSegment(p1, p2, _color);
-        debugDraw.drawSegment(x2, p2, _color);
-    }
-  }
+  void render(DebugDraw debugDraw) {}
 }