+++ packages/forge2d/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.11.0
+
+> Note: This release has breaking changes.
+
+ - **FEAT**: allowed specifying `BodyDef` properties via constructor (#51). ([aa3f5f7c](https://github.com/flame-engine/forge2d/commit/aa3f5f7cd604f1c7e3a8d90622dea812e3c63566))
+ - **FEAT**: allowed specifying `FixtureDef` properties via constructor (#52). ([1c09a139](https://github.com/flame-engine/forge2d/commit/1c09a1395deb97b53638791238ab094870a201b8))
+ - **BREAKING** **FEAT**: allow modifying gravity in the x-axis and y-axis (#53). ([c14b4245](https://github.com/flame-engine/forge2d/commit/c14b4245bd86f78418eb38a20ba0f0a62061af4f))
+ - **BREAKING** **FEAT**: made default friciton 0 (#54). ([e943a567](https://github.com/flame-engine/forge2d/commit/e943a567483510dcebd08bc8e5d4b53366670c67))
+
+## 0.10.0
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: moved Joint render method implementation (#49). ([b9128e5a](https://github.com/flame-engine/forge2d/commit/b9128e5a589140b596cf377faeaaf1140ada3c9e))
+ - **REFACTOR**: removed mock instances (#48). ([6fdf5646](https://github.com/flame-engine/forge2d/commit/6fdf5646a36b2749481269bfc56a0b88f35516c3))
+ - **BREAKING** **REFACTOR**: made `createJoint` use `Joint` over `JointDef` (#50). ([0e8f3655](https://github.com/flame-engine/forge2d/commit/0e8f3655b75d1c7c516b0ac79c7b6bde54fd3fcf))
+ - **BREAKING** **FEAT**: remove `JointType` (#45). ([42a39524](https://github.com/flame-engine/forge2d/commit/42a395241bb9c70932a1bbf6e84f936b52965d35))
+
 ## 0.9.0
 
 # CHANGELOG